--- src/controllers/index.js
@@ -1,3 +1,0 @@
-export const join = (req, res) => {
-  return res.send("JOIN");
-};
