+++ src/controllers/user/getChangePw.js
@@ -1,0 +1,5 @@
+export default (req, res) => {
+  return res.render("change-pw", {
+    pageTitle: "Change Password",
+  });
+};
