+++ src/controllers/user/getEdit.js
@@ -1,0 +1,5 @@
+export default (req, res) => {
+  return res.render("edit-profile", {
+    pageTitle: "Edit Profile",
+  });
+};
