+++ src/controllers/user/getJoin.js
@@ -1,0 +1,5 @@
+export default (req, res) => {
+  return res.render("join", {
+    pageTitle: "Join",
+  });
+};
