+++ src/controllers/user/getLogin.js
@@ -1,0 +1,5 @@
+export default (req, res) => {
+  return res.render("login", {
+    pageTitle: "Login",
+  });
+};
