+++ src/controllers/user/index.js
@@ -1,0 +1,29 @@
+import getEdit from "./getEdit";
+import postEdit from "./postEdit";
+import getLogin from "./getLogin";
+import postLogin from "./postLogin";
+import getJoin from "./getJoin";
+import postJoin from "./postJoin";
+import logout from "./logout";
+import remove from "./remove";
+import see from "./see";
+import startGithubLogin from "./startGithubLogin";
+import finishGithubLogin from "./finishGithubLogin";
+import getChangePw from "./getChangePw";
+import postChangePw from "./postChangePw";
+
+export {
+  getEdit,
+  postEdit,
+  getLogin,
+  postLogin,
+  getJoin,
+  postJoin,
+  logout,
+  remove,
+  see,
+  startGithubLogin,
+  finishGithubLogin,
+  getChangePw,
+  postChangePw,
+};
