+++ src/controllers/user/logout.js
@@ -1,0 +1,4 @@
+export default (req, res) => {
+  req.session.destroy();
+  return res.redirect("/");
+};
