+++ src/controllers/user/remove.js
@@ -1,0 +1,3 @@
+export default (req, res) => {
+  return res.send("REMOVE USER");
+};
