+++ src/controllers/user/see.js
@@ -1,0 +1,3 @@
+export default (req, res) => {
+  return res.send("SEE USER");
+};
