+++ src/controllers/video/deleteVideo.js
@@ -1,0 +1,7 @@
+import Video from "../../models/Video";
+
+export default async (req, res) => {
+  const { id } = req.params;
+  await Video.findByIdAndDelete(id);
+  return res.redirect("/");
+};
