+++ src/controllers/video/getUpload.js
@@ -1,0 +1,5 @@
+export default (req, res) => {
+  return res.render("upload", {
+    pageTitle: "Upload",
+  });
+};
