+++ src/controllers/video/index.js
@@ -1,0 +1,19 @@
+import getEdit from "./getEdit";
+import postEdit from "./postEdit";
+import getUpload from "./getUpload";
+import postUpload from "./postUpload";
+import home from "./home";
+import search from "./search";
+import watch from "./watch";
+import deleteVideo from "./deleteVideo";
+
+export {
+  getEdit,
+  postEdit,
+  getUpload,
+  postUpload,
+  home,
+  search,
+  watch,
+  deleteVideo,
+};
