+++ src/controllers/video/postUpload.js
@@ -1,0 +1,18 @@
+import Video from "../../models/Video";
+
+export default async (req, res) => {
+  const { title, description, hashtags } = req.body;
+  try {
+    await Video.create({
+      title,
+      description,
+      hashtags: Video.formatHashtags(hashtags),
+    });
+    return res.redirect("/");
+  } catch (error) {
+    return res.status(400).render("upload", {
+      pageTitle: "Upload",
+      errorMessage: error._message,
+    });
+  }
+};
