+++ src/middlewares/index.js
@@ -1,0 +1,5 @@
+import locals from "./locals.js";
+import protect from "./protect.js";
+import publicOnly from "./publicOnly.js";
+
+export { locals, protect, publicOnly };
