+++ src/middlewares/locals.js
@@ -1,6 +1,6 @@
 export default (req, res, next) => {
   res.locals.loggedIn = !!req.session.loggedIn;
   res.locals.siteName = "siteName";
-  res.locals.loggedInUser = req.session.user;
+  res.locals.loggedInUser = req.session.user || {};
   next();
 };
