+++ src/middlewares/protect.js
@@ -1,0 +1,7 @@
+export default (req, res, next) => {
+  if (req.session.loggedIn) {
+    next();
+  } else {
+    return res.redirect("/login");
+  }
+};
