+++ src/middlewares/publicOnly.js
@@ -1,0 +1,7 @@
+export default (req, res, next) => {
+  if (!req.session.loggedIn) {
+    return next();
+  } else {
+    return res.redirect("/");
+  }
+};
