+++ src/routers/user.js
@@ -1,18 +1,23 @@
 import express from "express";
 import {
   remove,
-  edit,
+  getEdit,
+  postEdit,
   see,
   startGithubLogin,
   finishGithubLogin,
+  getChangePw,
+  postChangePw,
 } from "../controllers/user";
+import { protect, publicOnly } from "../middlewares";
 
 const user = express.Router();
 
-user.get("/edit", edit);
+user.route("/edit").all(protect).get(getEdit).post(postEdit);
 user.get("/remove", remove);
-user.get("/:id", see);
-user.get("/github/start", startGithubLogin);
-user.get("/github/finish", finishGithubLogin);
+user.get("/:id([0-9a-f]{24})", see);
+user.route("/github/start").all(publicOnly).get(startGithubLogin);
+user.route("/github/finish").all(publicOnly).get(finishGithubLogin);
+user.route("/change-password").all(protect).get(getChangePw).post(postChangePw);
 
 export default user;
