+++ src/views/base.pug
@@ -6,7 +6,7 @@ html(lang="ko")
     body
         header
             h1=pageTitle
-            block header
+            include partials/nav
         main
             block content  
-    include partials/footer.pug
+    include partials/footer
