+++ src/views/change-pw.pug
@@ -1,0 +1,10 @@
+extends base
+
+block content 
+    if errorMsg
+        span=errorMsg
+    form(method="POST")
+        input(name="prevPw" placeholder="old password" type="password")
+        input(name="newPw" placeholder="new password" type="password")
+        input(name="confirmation" placeholder="new password confirmation" type="password")
+        input(type="submit" value="Change Password")
