+++ src/views/edit-profile.pug
@@ -1,0 +1,12 @@
+extends base
+
+block content 
+    form(method="POST")
+        input(name="name" type="text" required placeholder="name" value=loggedInUser.name)
+        input(name="email" type="email" required placeholder="email" value=loggedInUser.email)
+        input(name="userName" type="text" required placeholder="username" value=loggedInUser.userName)
+        input(name="location" type="text" placeholder="location" value=loggedInUser.location)
+        input(type="submit" value="Update Profile")
+        hr
+        if !loggedInUser.githubLoginOnly
+            a(href="change-password") Change Password &rarr;
