+++ src/views/edit.pug
@@ -1,7 +1,4 @@
-extends base.pug  
-
-block header
-    include partials/nav
+extends base 
 
 block content
     h4 Update Video
