+++ src/views/home.pug
@@ -1,8 +1,5 @@
-extends base.pug
+extends base
 include mixins/video          
-
-block header
-    include partials/nav
 
 block content
     each v in videos
