+++ src/views/login.pug
@@ -1,8 +1,5 @@
 extends base
 
-block header 
-    include partials/nav
-
 block content
     if errMsg
         span=errMsg
