+++ src/views/search.pug
@@ -1,9 +1,6 @@
 extends base
 include mixins/video
 
-block header
-    include partials/nav
-
 block content
     form(method="GET")
         input(name="keyword" placeholder="Search by title", type="text")
