+++ src/views/watch.pug
@@ -1,4 +1,4 @@
-extends base.pug
+extends base
 
 block content
     div
